third_party/blink/web_tests/flag-specific/highdpi/svg/as-list-image/svg-list-image-intrinsic-size-zoom-expected.txt

This is a testharness.js-based test.
PASS Zoom to 2 and list height should be equal to line-height
PASS Zoom to 3 and list height should be equal to line-height
PASS Zoom to 4 and list height should be equal to line-height
PASS Zoom to 5 and list height should be equal to line-height
FAIL Zoom to 0.5 and list height should be equal to line-height assert_approx_equals: expected 500 +/- 0.5 but got 498.66668701171875
PASS Zoom to 0.2 and list height should be equal to line-height
PASS Zoom to 1 and list height should be equal to line-height
Harness: the test ran to completion.
